# mechanic_shop_api/application/static/swagger.yaml
swagger: '2.0'
info:
  title: "Mechanic Shop API"
  description: "This is an API to manage a Mechanic Shop - tracking customers, mechanics, service tickets, and inventory"
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
paths:
  /customers/:
    post:
      tags:
        - Customers
      summary: "Create a new customer"
      description: "Register a new customer in the system. Email must be unique."
      parameters:
        - in: "body"
          name: "body"
          description: "Customer information for registration"
          required: true
          schema:
            $ref: "#/definitions/CreateCustomerPayload"
      responses:
        201:
          description: "Customer created successfully"
          schema:
            $ref: "#/definitions/CustomerResponse"
          examples:
            application/json:
              id: 1
              name: "John Doe"
              email: "john@email.com"
              phone: "123-456-7890"
              address: "123 Main St"
        400:
          description: "Validation error - missing or invalid fields"
          examples:
            application/json:
              email: ["Missing data for required field."]

    get:
      tags:
        - Customers
      summary: "Get all customers (paginated)"
      description: "Retrieve a paginated list of all customers. Supports page and per_page query parameters. This route is cached for 60 seconds."
      parameters:
        - in: "query"
          name: "page"
          type: "integer"
          description: "Page number (default: 1)"
          required: false
        - in: "query"
          name: "per_page"
          type: "integer"
          description: "Number of items per page (default: 10)"
          required: false
      responses:
        200:
          description: "Successfully retrieved customers"
          schema:
            $ref: "#/definitions/CustomersListResponse"
          examples:
            application/json:
              items:
                - id: 1
                  name: "John Doe"
                  email: "john@email.com"
                  phone: "123-456-7890"
                  address: "123 Main St"
                - id: 2
                  name: "Jane Smith"
                  email: "jane@email.com"
                  phone: "987-654-3210"
                  address: "456 Oak Ave"
              page: 1
              per_page: 10
              total: 50
              pages: 5
  /customers/{customer_id}:
    get:
      tags:
        - Customers
      summary: "Get a single customer by ID"
      description: "Retrieve detailed information about a specific customer using their ID."
      parameters:
        - in: "path"
          name: "customer_id"
          type: "integer"
          required: true
          description: "The ID of the customer to retrieve"
      responses:
        200:
          description: "Successfully retrieved customer"
          schema:
            $ref: "#/definitions/CustomerResponse"
          examples:
            application/json:
              id: 1
              name: "John Doe"
              email: "john@email.com"
              phone: "123-456-7890"
              address: "123 Main St"
        404:
          description: "Customer not found"
          examples:
            application/json:
              error: "Customer not found."

    put:
      tags:
        - Customers
      summary: "Update customer information"
      description: "Update a customer's information. Requires authentication - customers can only update their own account."
      security:
        - bearerAuth: []
      parameters:
        - in: "path"
          name: "customer_id"
          type: "integer"
          required: true
          description: "The ID of the customer to update"
        - in: "body"
          name: "body"
          description: "Updated customer information (all fields optional)"
          required: true
          schema:
            $ref: "#/definitions/UpdateCustomerPayload"
      responses:
        200:
          description: "Customer updated successfully"
          schema:
            $ref: "#/definitions/CustomerResponse"
          examples:
            application/json:
              id: 1
              name: "John Updated"
              email: "john@email.com"
              phone: "123-456-7890"
              address: "456 New St"
        400:
          description: "Validation error"
        401:
          description: "Missing or invalid token"
          examples:
            application/json:
              error: "Token is missing or invalid"
        403:
          description: "Unauthorized - trying to update another customer's account"
          examples:
            application/json:
              error: "Unauthorized to update this customer."
        404:
          description: "Customer not found"
          examples:
            application/json:
              error: "Customer not found."

    delete:
      tags:
        - Customers
      summary: "Delete a customer account"
      description: "Delete a customer from the system. Requires authentication - customers can only delete their own account."
      security:
        - bearerAuth: []
      parameters:
        - in: "path"
          name: "customer_id"
          type: "integer"
          required: true
          description: "The ID of the customer to delete"
      responses:
        200:
          description: "Customer deleted successfully"
          schema:
            $ref: "#/definitions/DeleteResponse"
          examples:
            application/json:
              message: "Customer 1 deleted."
        401:
          description: "Missing or invalid token"
          examples:
            application/json:
              error: "Token is missing or invalid"
        403:
          description: "Unauthorized - trying to delete another customer's account"
          examples:
            application/json:
              error: "Unauthorized to delete this customer."
        404:
          description: "Customer not found"
          examples:
            application/json:
              error: "Customer not found."
  /customers/login:
    post:
      tags:
        - Customers
      summary: "Customer login endpoint (generates token)"
      description: "Validates customer credentials and returns a JWT token for authentication. Rate limited to 5 requests per minute."
      parameters:
        - in: "body"
          name: "body"
          description: "Customer login credentials"
          required: true
          schema:
            $ref: "#/definitions/LoginCredentials"
      responses:
        200:
          description: "Login successful - returns JWT token"
          schema:
            $ref: "#/definitions/LoginResponse"
          examples:
            application/json:
              token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
        401:
          description: "Invalid credentials"
          examples:
            application/json:
              error: "Invalid credentials."
        400:
          description: "Validation error - missing or invalid fields"
          examples:
            application/json:
              email: ["Missing data for required field."]
        429:
          description: "Rate limit exceeded - too many login attempts"
          examples:
            application/json:
              error: "Too many requests"
  /customers/my-tickets:
    get:
      tags:
        - Customers
      summary: "Get tickets for logged-in customer"
      description: "Retrieve all service tickets belonging to the authenticated customer. Requires valid JWT token."
      security:
        - bearerAuth: []
      responses:
        200:
          description: "Successfully retrieved customer's tickets"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ServiceTicketResponse"
          examples:
            application/json:
              - id: 1
                description: "Oil change and brake repair"
                vin: "1HGCM82633A123456"
                customer_id: 1
                mechanics: []
                inventory_items: []
              - id: 2
                description: "Transmission service"
                vin: "1HGCM82633A123456"
                customer_id: 1
                mechanics: []
                inventory_items: []
        401:
          description: "Missing or invalid token"
          examples:
            application/json:
              error: "Token is missing or invalid"
...
definitions:
  CreateCustomerPayload:
    type: "object"
    properties:
      name:
        type: "string"
        description: "Customer's full name"
      email:
        type: "string"
        description: "Customer's email address (must be unique)"
      password:
        type: "string"
        description: "Customer's password (will be hashed before storage)"
      phone:
        type: "string"
        description: "Customer's phone number (optional)"
      address:
        type: "string"
        description: "Customer's address (optional)"
    required:
      - name
      - email
      - password
  UpdateCustomerPayload:
    type: "object"
    properties:
      name:
        type: "string"
        description: "Updated customer name"
      email:
        type: "string"
        description: "Updated email address"
      password:
        type: "string"
        description: "Updated password (will be hashed)"
      phone:
        type: "string"
        description: "Updated phone number"
      address:
        type: "string"
        description: "Updated address"
  CustomerResponse:
    type: "object"
    properties:
      id:
        type: "integer"
        description: "Customer's unique ID"
      name:
        type: "string"
        description: "Customer's full name"
      email:
        type: "string"
        description: "Customer's email address"
      phone:
        type: "string"
        description: "Customer's phone number"
      address:
        type: "string"
        description: "Customer's address"
  CustomersListResponse:
    type: "object"
    properties:
      items:
        type: "array"
        description: "Array of customer objects"
        items:
          $ref: "#/definitions/CustomerResponse"
      page:
        type: "integer"
        description: "Current page number"
      per_page:
        type: "integer"
        description: "Number of items per page"
      total:
        type: "integer"
        description: "Total number of customers in database"
      pages:
        type: "integer"
        description: "Total number of pages available"
  LoginCredentials:
    type: "object"
    properties:
      email:
        type: "string"
        description: "Customer's email address"
      password:
        type: "string"
        description: "Customer's password"
    required:
      - email
      - password
  LoginResponse:
    type: "object"
    properties:
      token:
        type: "string"
        description: "JWT authentication token to use in Authorization header for protected routes"
  DeleteResponse:
    type: "object"
    properties:
      message:
        type: "string"
        description: "Confirmation message that the resource was deleted"
...
  MechanicResponse:
    type: "object"
    properties:
      id:
        type: "integer"
        description: "Mechanic's unique ID"
      name:
        type: "string"
        description: "Mechanic's full name"
      email:
        type: "string"
        description: "Mechanic's email address"
      phone:
        type: "string"
        description: "Mechanic's phone number"
      address:
        type: "string"
        description: "Mechanic's address"
      salary:
        type: "number"
        format: "float"
        description: "Mechanic's salary"
...
  ServiceTicketResponse:
    type: "object"
    properties:
      id:
        type: "integer"
        description: "Service ticket's unique ID"
      description:
        type: "string"
        description: "Description of the service"
      vin:
        type: "string"
        description: "Vehicle Identification Number"
      customer_id:
        type: "integer"
        description: "ID of the customer"
      mechanics:
        type: "array"
        description: "List of mechanics assigned to this ticket"
        items:
          $ref: "#/definitions/MechanicResponse"
      inventory_items:
        type: "array"
        description: "List of parts/items used in this service"
        items:
          $ref: "#/definitions/InventoryResponse"
...
  InventoryResponse:
    type: "object"
    properties:
      id:
        type: "integer"
        description: "Inventory item's unique ID"
      name:
        type: "string"
        description: "Name of the part/item"
      price:
        type: "number"
        format: "float"
        description: "Price of the part/item"